+++ test/unit/fuzz-tools.test.js
@@ -194,6 +194,7 @@ import * as hepatologyGibleedV201 from '../../lib/hepatology-gibleed-v201.js';
 import * as cvriskEnginesV202 from '../../lib/cvrisk-engines-v202.js';
 import * as periopFrailtyV203 from '../../lib/periop-frailty-v203.js';
 import * as nephroFluidsV204 from '../../lib/nephro-fluids-v204.js';
+import * as pulmCopdV205 from '../../lib/pulm-copd-v205.js';
 
 const MODULES = {
   'clinical.js': clinical,
@@ -332,6 +333,7 @@ const MODULES = {
   'cvrisk-engines-v202.js': cvriskEnginesV202,
   'periop-frailty-v203.js': periopFrailtyV203,
   'nephro-fluids-v204.js': nephroFluidsV204,
+  'pulm-copd-v205.js': pulmCopdV205,
 };
 
 const MATRIX = [0, -1, 1e9, NaN, Infinity, -Infinity, '', undefined, null];